Library App
Created a Library app with two buttons. one of the button is to add Books
while the other button is used to view stored books.  The add book button
is linked to a form where user can register a book by filling the form.
on completing and submitting the form the book details is passed to the
registered book is displayed.
On clicking of the submit a function is called where the form details are
stored to local storage using the method localStorage.setItems where a
placeholder and the item details are set.
After that the stored items are retrieved from the localStorage using
localStorage.getItem('placeholder') being passed to the registered form by
passing the value as innerHTML on each of the table data using their ID